Ledger project namespace refactoring
