+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/App.cs
@@ -1,4 +1,4 @@
-﻿namespace Ledger.Net
+﻿namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet
 {
     public enum App
     {
