+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/BitcoinAddressType.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Ledger.Net
+﻿namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet
 {
     [Flags]
     public enum BitcoinAddressType
