+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/Constants.cs
@@ -1,4 +1,4 @@
-﻿namespace Ledger.Net
+﻿namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet
 {
     public static class Constants
     {
