+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/Exceptions/IncorrectLengthException.cs
@@ -1,4 +1,4 @@
-﻿namespace Ledger.Net.Exceptions
+﻿namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Exceptions
 {
     public class IncorrectLengthException : ResponseBaseException
     {
