+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/Exceptions/InstructionNotSupportedException.cs
@@ -1,4 +1,4 @@
-﻿namespace Ledger.Net.Exceptions
+﻿namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Exceptions
 {
     public class InstructionNotSupportedException : ResponseBaseException
     {
