+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/Helpers.cs
@@ -1,10 +1,8 @@
 ﻿using Hardwarewallets.Net.Model;
-using Ledger.Net.Exceptions;
-using Ledger.Net.Responses;
-using System;
-using System.IO;
+using Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Exceptions;
+using Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Responses.Abstract;
 
-namespace Ledger.Net
+namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet
 {
     public static class Helpers
     {
