+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/ICoinInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace Ledger.Net
+﻿namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet
 {
     public interface ICoinInfo
     {
