+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/IHandlesRequest.cs
@@ -1,8 +1,7 @@
-﻿using Ledger.Net.Requests;
-using Ledger.Net.Responses;
-using System.Threading.Tasks;
+﻿using Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Requests.Abstract;
+using Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Responses.Abstract;
 
-namespace Ledger.Net
+namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet
 {
     public interface IHandlesRequest
     {
