+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/LedgerManagerFactory.cs
@@ -1,6 +1,6 @@
 ﻿using Device.Net;
 
-namespace Ledger.Net
+namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet
 {
     public class LedgerManagerFactory : ILedgerManagerFactory
     {
