+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/Responses/Abstract/ResponseBase.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Ledger.Net.Responses
+﻿namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Responses.Abstract
 {
     public abstract class ResponseBase
     {
