+++ Zilliqa.DesktopWallet.Device.Ledger/LedgerNet/Responses/Concrete/TronAppSignatureResponse.cs
@@ -1,4 +1,6 @@
-﻿namespace Ledger.Net.Responses
+﻿using Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Responses.Abstract;
+
+namespace Zilliqa.DesktopWallet.Device.Ledger.LedgerNet.Responses.Concrete
 {
     public class TronAppSignatureResponse : ResponseBase
     {
