+++ Zilliqa.DesktopWallet.Device.Ledger/ZilliqaCoinUtility.cs
@@ -1,6 +1,6 @@
-﻿using System;
+﻿using Zilliqa.DesktopWallet.Device.Ledger.LedgerNet;
 
-namespace Ledger.Net
+namespace Zilliqa.DesktopWallet.Device.Ledger
 {
     public class ZilliqaCoinUtility : ICoinUtility
     {
